Renames reporter hook
