--- src/signals/incident-management/containers/ReporterContainer/__tests__/hook.test.ts
@@ -1,4 +1,0 @@
-// Copyright (C) 2021 Gemeente Amsterdam
-describe('Reporter hook', () => {
-  it.todo('Test reporter hook')
-})
